+++ src/pharmpy/tools/__init__.py
@@ -18,6 +18,7 @@ __all__ = (
     'run_modelsearch',  # pyright: ignore [reportUnsupportedDunderAll]
     'run_ruvsearch',  # pyright: ignore [reportUnsupportedDunderAll]
     'run_tool',  # pyright: ignore [reportUnsupportedDunderAll]
+    'summarize_errors',  # pyright: ignore [reportUnsupportedDunderAll]
     'summarize_individuals',  # pyright: ignore [reportUnsupportedDunderAll]
     'summarize_individuals_count_table',  # pyright: ignore [reportUnsupportedDunderAll]
     'write_results',  # pyright: ignore [reportUnsupportedDunderAll]
@@ -34,6 +35,7 @@ _run_keys = {
     'retrieve_final_model',
     'retrieve_models',
     'run_tool',
+    'summarize_errors',
     'write_results',
 }