+++ src/pharmpy/tools/common.py
@@ -3,7 +3,8 @@ from typing import Any
 
 from pharmpy.deps import numpy as np
 from pharmpy.deps import pandas as pd
-from pharmpy.modeling import rank_models, summarize_errors, summarize_modelfit_results, update_inits
+from pharmpy.modeling import rank_models, summarize_modelfit_results, update_inits
+from pharmpy.tools import summarize_errors
 
 from .funcs import summarize_individuals, summarize_individuals_count_table