+++ tests/tools/test_run.py
@@ -1,6 +1,7 @@
 import importlib
 import inspect
 import os
+import shutil
 from pathlib import Path
 
 import pytest
@@ -13,6 +14,7 @@ from pharmpy.tools.run import (
     _get_run_setup,
     retrieve_final_model,
     retrieve_models,
+    summarize_errors,
 )
 from pharmpy.utils import TemporaryDirectoryChanger
 from pharmpy.workflows import LocalDirectoryToolDatabase, local_dask
@@ -174,3 +176,33 @@ def test_retrieve_final_model(testdata):
     res = read_results(results_json_none)
     with pytest.raises(ValueError, match='Attribute \'final_model_name\' is None'):
         retrieve_final_model(res)
+
+
+def test_summarize_errors(load_model_for_test, testdata, tmp_path, pheno_path):
+    with TemporaryDirectoryChanger(tmp_path):
+        model = load_model_for_test(pheno_path)
+        shutil.copy2(testdata / 'pheno_data.csv', tmp_path)
+
+        error_path = testdata / 'nonmem' / 'errors'
+
+        shutil.copy2(testdata / 'nonmem' / 'pheno_real.mod', tmp_path / 'pheno_no_header.mod')
+        shutil.copy2(error_path / 'no_header_error.lst', tmp_path / 'pheno_no_header.lst')
+        shutil.copy2(testdata / 'nonmem' / 'pheno_real.ext', tmp_path / 'pheno_no_header.ext')
+        model_no_header = load_model_for_test('pheno_no_header.mod')
+        model_no_header.datainfo = model_no_header.datainfo.derive(path=tmp_path / 'pheno_data.csv')
+
+        shutil.copy2(testdata / 'nonmem' / 'pheno_real.mod', tmp_path / 'pheno_rounding_error.mod')
+        shutil.copy2(error_path / 'rounding_error.lst', tmp_path / 'pheno_rounding_error.lst')
+        shutil.copy2(testdata / 'nonmem' / 'pheno_real.ext', tmp_path / 'pheno_rounding_error.ext')
+        model_rounding_error = load_model_for_test('pheno_rounding_error.mod')
+        model_rounding_error.datainfo = model_rounding_error.datainfo.derive(
+            path=tmp_path / 'pheno_data.csv'
+        )
+
+        models = [model, model_no_header, model_rounding_error]
+        summary = summarize_errors(models)
+
+        assert 'pheno_real' not in summary.index.get_level_values('model')
+        assert len(summary.loc[('pheno_no_header', 'WARNING')]) == 1
+        assert len(summary.loc[('pheno_no_header', 'ERROR')]) == 2
+        assert len(summary.loc[('pheno_rounding_error', 'ERROR')]) == 2